+++ layers/CMSSchema/packages/pages/src/ConditionalOnModule/Comments/FieldResolvers/ObjectType/CommentablePageObjectTypeFieldResolver.php
@@ -1,0 +1,40 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPCMSSchema\Pages\ConditionalOnModule\Comments\FieldResolvers\ObjectType;
+
+use PoPCMSSchema\Comments\FieldResolvers\ObjectType\AbstractCommentableCustomPostObjectTypeFieldResolver;
+use PoPCMSSchema\Pages\TypeAPIs\PageTypeAPIInterface;
+use PoPCMSSchema\Pages\TypeResolvers\ObjectType\PageObjectTypeResolver;
+use PoP\ComponentModel\TypeResolvers\ObjectType\ObjectTypeResolverInterface;
+
+class CommentablePageObjectTypeFieldResolver extends AbstractCommentableCustomPostObjectTypeFieldResolver
+{
+    private ?PageTypeAPIInterface $pageTypeAPI = null;
+
+    final public function setPageTypeAPI(PageTypeAPIInterface $pageTypeAPI): void
+    {
+        $this->pageTypeAPI = $pageTypeAPI;
+    }
+    final protected function getPageTypeAPI(): PageTypeAPIInterface
+    {
+        /** @var PageTypeAPIInterface */
+        return $this->pageTypeAPI ??= $this->instanceManager->getInstance(PageTypeAPIInterface::class);
+    }
+
+    /**
+     * @return array<class-string<ObjectTypeResolverInterface>>
+     */
+    public function getObjectTypeResolverClassesToAttachTo(): array
+    {
+        return [
+            PageObjectTypeResolver::class,
+        ];
+    }
+
+    protected function getCustomPostType(): string
+    {
+        return $this->getPageTypeAPI()->getPageCustomPostType();
+    }
+}