+++ layers/CMSSchema/packages/pages/src/Module.php
@@ -8,6 +8,7 @@ use PoP\Root\Module\ModuleInterface;
 use PoP\Root\App;
 use PoPAPI\API\Module as APIModule;
 use PoPAPI\RESTAPI\Module as RESTAPIModule;
+use PoPCMSSchema\Comments\Module as CommentsModule;
 use PoP\Root\Module\AbstractModule;
 
 class Module extends AbstractModule
@@ -35,6 +36,7 @@ class Module extends AbstractModule
         return [
             \PoPAPI\API\Module::class,
             \PoPAPI\RESTAPI\Module::class,
+            \PoPCMSSchema\Comments\Module::class,
         ];
     }
 
@@ -54,6 +56,13 @@ class Module extends AbstractModule
         }
         if (class_exists(RESTAPIModule::class) && App::getModule(RESTAPIModule::class)->isEnabled()) {
             $this->initServices(dirname(__DIR__), '/ConditionalOnModule/RESTAPI');
+        }
+        if (class_exists(CommentsModule::class) && App::getModule(CommentsModule::class)->isEnabled()) {
+            $this->initSchemaServices(
+                dirname(__DIR__),
+                $skipSchema || in_array(CommentsModule::class, $skipSchemaModuleClasses),
+                '/ConditionalOnModule/Comments'
+            );
         }
     }
 }