+++ layers/GraphQLAPIForWP/phpunit-packages/graphql-api-for-wp/tests/Unit/Faker/EnabledCommentableCustomPostWPFakerFixtureQueryExecutionGraphQLServerTest.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PHPUnitForGraphQLAPI\GraphQLAPI\Unit\Faker;
+
+use GraphQLByPoP\GraphQLServer\Unit\EnabledFixtureQueryExecutionGraphQLServerTestCaseTrait;
+
+class EnabledCommentableCustomPostWPFakerFixtureQueryExecutionGraphQLServerTest extends AbstractCommentableCustomPostWPFakerFixtureQueryExecutionGraphQLServerTest
+{
+    use EnabledFixtureQueryExecutionGraphQLServerTestCaseTrait;
+}